Pull reginfo data for monthly ES rules and re-organize modules
