+++ data/cumulative_es_rules/cumulative_econ_significant_rules_by_presidential_month.csv
@@ -95,4 +95,4 @@ Sep,92,151,,320,310,448,,,
 Oct,93,153,,323,323,457,,,
 Nov,94,155,,333,338,468,,,
 Dec,95,155,,339,347,482,,,
-Jan,96,159,,361,358,497,,,
+Jan,96,159,,361,358,496,,,
