+++ data/monthly_es_rules/monthly_significant_rules_by_admin.csv
@@ -43,93 +43,6 @@ Biden,2024,Jun,41,7.0,1.0,6.0
 Biden,2024,Jul,42,11.0,1.0,10.0
 Biden,2024,Aug,43,21.0,9.0,12.0
 Biden,2024,Sep,44,10.0,4.0,6.0
-Biden,2021,Feb,1,9.0,6.0,3.0
-Biden,2021,Mar,2,11.0,4.0,7.0
-Biden,2021,Apr,3,9.0,2.0,7.0
-Biden,2021,May,4,16.0,10.0,6.0
-Biden,2021,Jun,5,12.0,2.0,10.0
-Biden,2021,Jul,6,12.0,8.0,4.0
-Biden,2021,Aug,7,21.0,10.0,11.0
-Biden,2021,Sep,8,12.0,3.0,9.0
-Biden,2021,Oct,9,15.0,6.0,9.0
-Biden,2021,Nov,10,15.0,10.0,5.0
-Biden,2021,Dec,11,23.0,5.0,18.0
-Biden,2022,Jan,12,9.0,3.0,6.0
-Biden,2022,Feb,13,4.0,2.0,2.0
-Biden,2022,Mar,14,15.0,2.0,13.0
-Biden,2022,Apr,15,12.0,3.0,9.0
-Biden,2022,May,16,16.0,8.0,8.0
-Biden,2022,Jun,17,12.0,0.0,12.0
-Biden,2022,Jul,18,11.0,4.0,7.0
-Biden,2022,Aug,19,14.0,5.0,9.0
-Biden,2022,Sep,20,12.0,4.0,8.0
-Biden,2022,Oct,21,10.0,6.0,4.0
-Biden,2022,Nov,22,14.0,7.0,7.0
-Biden,2022,Dec,23,11.0,4.0,7.0
-Biden,2023,Jan,24,13.0,4.0,9.0
-Biden,2023,Feb,25,12.0,2.0,10.0
-Biden,2023,Mar,26,13.0,1.0,12.0
-Biden,2023,Apr,27,10.0,3.0,7.0
-Biden,2023,May,28,11.0,3.0,8.0
-Biden,2023,Jun,29,14.0,4.0,10.0
-Biden,2023,Jul,30,12.0,5.0,7.0
-Biden,2023,Aug,31,18.0,12.0,6.0
-Biden,2023,Sep,32,14.0,12.0,2.0
-Biden,2023,Oct,33,17.0,9.0,8.0
-Biden,2023,Nov,34,14.0,10.0,4.0
-Biden,2023,Dec,35,20.0,8.0,12.0
-Biden,2024,Jan,36,11.0,9.0,2.0
-Biden,2024,Feb,37,20.0,8.0,12.0
-Biden,2024,Mar,38,27.0,8.0,19.0
-Biden,2024,Apr,39,66.0,34.0,32.0
-Biden,2024,May,40,48.0,27.0,21.0
-Biden,2024,Jun,41,7.0,1.0,6.0
-Biden,2024,Jul,42,11.0,1.0,10.0
-Biden,2024,Aug,43,21.0,9.0,12.0
-Biden,2021,Feb,1,9.0,6.0,3.0
-Biden,2021,Mar,2,11.0,4.0,7.0
-Biden,2021,Apr,3,9.0,2.0,7.0
-Biden,2021,May,4,16.0,10.0,6.0
-Biden,2021,Jun,5,12.0,2.0,10.0
-Biden,2021,Jul,6,12.0,8.0,4.0
-Biden,2021,Aug,7,21.0,10.0,11.0
-Biden,2021,Sep,8,12.0,3.0,9.0
-Biden,2021,Oct,9,15.0,6.0,9.0
-Biden,2021,Nov,10,15.0,10.0,5.0
-Biden,2021,Dec,11,23.0,5.0,18.0
-Biden,2022,Jan,12,9.0,3.0,6.0
-Biden,2022,Feb,13,4.0,2.0,2.0
-Biden,2022,Mar,14,15.0,2.0,13.0
-Biden,2022,Apr,15,12.0,3.0,9.0
-Biden,2022,May,16,16.0,8.0,8.0
-Biden,2022,Jun,17,12.0,0.0,12.0
-Biden,2022,Jul,18,11.0,4.0,7.0
-Biden,2022,Aug,19,14.0,5.0,9.0
-Biden,2022,Sep,20,12.0,4.0,8.0
-Biden,2022,Oct,21,10.0,6.0,4.0
-Biden,2022,Nov,22,14.0,7.0,7.0
-Biden,2022,Dec,23,11.0,4.0,7.0
-Biden,2023,Jan,24,13.0,4.0,9.0
-Biden,2023,Feb,25,12.0,2.0,10.0
-Biden,2023,Mar,26,13.0,1.0,12.0
-Biden,2023,Apr,27,10.0,3.0,7.0
-Biden,2023,May,28,11.0,3.0,8.0
-Biden,2023,Jun,29,14.0,4.0,10.0
-Biden,2023,Jul,30,12.0,5.0,7.0
-Biden,2023,Aug,31,18.0,12.0,6.0
-Biden,2023,Sep,32,14.0,12.0,2.0
-Biden,2023,Oct,33,17.0,9.0,8.0
-Biden,2023,Nov,34,14.0,10.0,4.0
-Biden,2023,Dec,35,20.0,8.0,12.0
-Biden,2024,Jan,36,11.0,9.0,2.0
-Biden,2024,Feb,37,20.0,8.0,12.0
-Biden,2024,Mar,38,27.0,8.0,19.0
-Biden,2024,Apr,39,66.0,34.0,32.0
-Biden,2024,May,40,48.0,27.0,21.0
-Biden,2024,Jun,41,7.0,1.0,6.0
-Biden,2024,Jul,42,11.0,1.0,10.0
-Biden,2024,Aug,43,21.0,9.0,12.0
-Biden,2024,Sep,44,10.0,4.0,6.0
 Biden,2024,Oct,45,16.0,4.0,12.0
 Biden,2024,Nov,46,22.0,9.0,13.0
 Biden,2024,Dec,47,39.0,9.0,30.0
