+++ data/monthly_es_rules/update_monthly_es_rules.py
@@ -1,4 +1,5 @@
-print('NOTE: The current code can only update data for the Biden and following administrations.')
+#----------------------Code to Update Monthly Economically Significant Rules by Presidential Month---------------------
+#-----------------------------------------Author: Zhoudan Xie----------------------------------------------------------
 
 #%% Import packages
 import os
@@ -7,13 +8,17 @@ import sys
 # Import customized functions
 dir_path=os.path.dirname(os.path.realpath(__file__))
 sys.path.insert(0, f'{dir_path}/../py_funcs')
-import frcount
+import update_admin
 
 #%% Specify the file path of the current dataset
 file_path=f'{dir_path}/monthly_econ_significant_rules_by_presidential_month.csv'
 
+#%% Specify rule type and data type
+rule_type='es'
+data_type='monthly'
+
 #%% Run functions to update the data
-df=frcount.main(dir_path,file_path,rule_type='es',type='monthly')
+df=update_admin.main(dir_path,file_path,rule_type,data_type)
 
 #%% Export
 df.to_csv(file_path, index=False)
